pyproject.toml
[project]
name = "drf-porject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.1",
    "psycopg>=3.3.3",
]
